Modify DirectControl for v5.0.0
